Refactor project summary data aggregation
Moved comment and file count logic from ProjectRepository to dedicated ProjectCommentService and ProjectFileService. Updated ScheduleRepository to provide next schedule date. Adjusted ProjectSummaryDTO to use Integer for counts. Removed unused ProjectPageDTO. Improved separation of concerns and modularity in project-related services.

+++ src/main/java/com/solarize/solarizeWebBackend/modules/project/dto/ProjectSummaryDTO.java
@@ -16,16 +16,12 @@ import java.time.LocalDateTime;
 @Getter
 @Setter
 public class ProjectSummaryDTO {
-
     private Long id;
     private String projectTitle;
     private String status;
     private LocalDateTime nextSchedule;
     private CoworkerResponseDto responsible;
     private ClientResponseDTO client;
-    private Long commentCount;
-    private Long fileCount;
-
-
-
+    private Integer commentCount;
+    private Integer fileCount;
 }
